Move tests for `MoneyCalculator` to the `model` module.
